+++ miasm2/analysis/binary.py
@@ -2,7 +2,7 @@ import logging
 
 from miasm2.core.bin_stream import bin_stream_str, bin_stream_elf, bin_stream_pe
 from miasm2.jitter.csts import PAGE_READ
-from miasm2.core.asmbloc import AsmSymbolPool
+from miasm2.core.asmblock import AsmSymbolPool
 
 
 log = logging.getLogger("binary")
