+++ miasm2/jitter/jitcore_cc_base.py
@@ -113,7 +113,7 @@ class JitCore_Cc_Base(JitCore):
     def hash_block(self, block):
         """
         Build a hash of the block @block
-        @block: asmbloc
+        @block: asmblock
         """
         block_raw = "".join(line.b for line in block.lines)
         block_hash = md5("%X_%s_%s_%s" % (block.label.offset,
